adding cookie and express session

+++ public/assignments/server/services/user.service.server.js
@@ -1,8 +1,13 @@
 /**
  * Created by hriya on 3/15/16.
  */
+var passport         = require('passport');
+var LocalStrategy    = require('passport-local').Strategy;
+
 module.exports = function (app, model) {
     "use strict";
+    var auth = authorized;
+
     app.post('/api/assignment/user', createUser);
     app.get('/api/assignment/user/:id', getUserById);
     app.all('/api/assignment/user', function (req, res, next) {
@@ -15,20 +20,120 @@ module.exports = function (app, model) {
             next();
         }
     });
-    app.get('/api/assignment/user', getAllUsers);
+    app.get('/api/assignment/user', auth,getAllUsers);
 
-    app.put('/api/assignment/user/:userId', updateUser);
-    app.delete('/api/assignment/user/:id', deleteUser);
+    app.put('/api/assignment/user/:userId',auth, updateUser);
+    app.delete('/api/assignment/user/:id',auth, deleteUser);
+    app.post('/api/login',passport.authenticate('local'), login);
+    app.post('/api/logout',         logout);
+    app.post('/api/register',       register);
+    app.post('/api/user',     auth, createUser);
+    app.get ('/api/loggedin',       loggedin);
+
+    passport.use(new LocalStrategy(localStrategy));
+    passport.serializeUser(serializeUser);
+    passport.deserializeUser(deserializeUser);
+
+
+    function localStrategy(username, password, done) {
+        model
+            .findUserByCredentials(username, password)
+            .then(
+                function(user) {
+                    if (!user) { return done(null, false); }
+                    return done(null, user);
+                },
+                function(err) {
+                    if (err) { return done(err); }
+                }
+            );
+    }
+
+    function serializeUser(user, done) {
+        done(null, user);
+    }
+
+    function deserializeUser(user, done) {
+        model
+            .findUserById(user._id)
+            .then(
+                function(user){
+                    done(null, user);
+                },
+                function(err){
+                    done(err, null);
+                }
+            );
+    }
+
+    function login(req, res) {
+        var user = req.user;
+        res.json(user);
+    }
+
+    function loggedin(req, res) {
+        res.send(req.isAuthenticated() ? req.user : '0');
+    }
+
+    function logout(req, res) {
+        req.logOut();
+        res.send(200);
+    }
+
+    function register(req, res) {
+        var newUser = req.body;
+        newUser.roles = ['user'];
+
+        model
+            .findUserByUsername(newUser.username)
+            .then(
+                function(user){
+                    if(user) {
+                        res.json(null);
+                    } else {
+                        model.createUser(newUser).then(
+                            function(user){
+                                if(user){
+                                    req.login(user, function(err) {
+                                        if(err) {
+                                            res.status(400).send(err);
+                                        } else {
+                                            res.json(user);
+                                        }
+                                    });
+                                }
+                            },
+                            function(err){
+                                res.status(400).send(err);
+                            }
+                        );
+                    }
+                },
+                function(err){
+                    res.status(400).send(err);
+                }
+            );
+
+    }
 
 
     function createUser(req, res) {
-        var user = req.body;
-        model.createUser(user).then(
+        var newUser = req.body;
+        delete newUser['_id'];
+        if(newUser.roles && newUser.roles.length > 1) {
+            newUser.roles = newUser.roles.split(",");
+        } else {
+
+            newUser.roles = ["user"];
+        }
+
+        model.createUser(newUser).then(
             function (doc) {
 
                 res.json(doc);
             },
             function (err) {
+                console.log(err);
                 res.status(400).send(err);
             }
         );
@@ -40,7 +145,9 @@ module.exports = function (app, model) {
     }
 
     function getAllUsers(req, res) {
-        model.findAllUsers().then(
+        if(isAdmin(req.user)) {
+
+            model.findAllUsers().then(
             function (doc) {
                 res.json(doc);
             },
@@ -48,6 +155,9 @@ module.exports = function (app, model) {
                 res.status(400).send(err);
             }
         );
+        } else {
+            res.status(403);
+        }
     }
 
 
@@ -74,6 +184,12 @@ module.exports = function (app, model) {
     function updateUser(req, res) {
         var id = req.params.userId;
         var user = req.body;
+        if(!isAdmin(req.user)) {
+            delete user.roles;
+        }
+        if(typeof user.roles == "string") {
+            user.roles = user.roles.split(",");
+        }
 
         model.updateUser(id, user).then(
             function (doc) {
@@ -87,7 +203,10 @@ module.exports = function (app, model) {
 
     function deleteUser(req, res) {
         var id = req.params.id;
-        model.deleteUserById(id).then(
+        //if same user as logged in do not delete
+        if(isAdmin(req.user)) {
+
+            model.deleteUserById(id).then(
             function (doc) {
                 res.json(doc);
             },
@@ -95,5 +214,23 @@ module.exports = function (app, model) {
                 res.status(400).send(err);
             }
         );
+        } else {
+            res.status(403);
+        }
     }
+
+    function isAdmin(user) {
+        if(user.roles.indexOf("admin") > -1) {
+            return true
+        }
+        return false;
+    }
+
+    function authorized (req, res, next) {
+        if (!req.isAuthenticated()) {
+            res.send(401);
+        } else {
+            next();
+        }
+    };
 }
